% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_local.R
\name{as_loa_workbook}
\alias{as_loa_workbook}
\title{Accept a List of Analysis as a Path or an Already-Read Workbook}
\usage{
as_loa_workbook(loa)
}
\arguments{
\item{loa}{A path, or the list from \code{\link[=read_loa_workbook]{read_loa_workbook()}}.}
}
\value{
A workbook list.
}
\description{
Accept a List of Analysis as a Path or an Already-Read Workbook
}
\keyword{internal}
