% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_xlsx.R
\name{ck_add_rect}
\alias{ck_add_rect}
\title{Queue a Rectangle of Cells for a Style}
\usage{
ck_add_rect(sg, key, rows, cols)
}
\arguments{
\item{sg}{A style grid from \code{\link[=ck_new_style_grid]{ck_new_style_grid()}}.}

\item{key}{A style key returned by \code{\link[=ck_style]{ck_style()}}.}

\item{rows}{Integer vector of sheet rows.}

\item{cols}{Integer vector of sheet columns.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Queue a Rectangle of Cells for a Style
}
\keyword{internal}
