% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{ck_analysis_type_labels}
\alias{ck_analysis_type_labels}
\title{Default Analysis Type Labels}
\usage{
ck_analysis_type_labels()
}
\value{
A dataframe with \code{analysis_type} and \code{label_analysis_type}.
}
\description{
Default Analysis Type Labels
}
\keyword{internal}
