% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_xlsx.R
\name{ck_border}
\alias{ck_border}
\title{Build a Border Specification}
\usage{
ck_border(
  sides = c("Top", "Bottom", "Left", "Right"),
  thick = character(0),
  thin_colour = "#AFDFE4",
  thick_colour = "#003D58"
)
}
\arguments{
\item{sides}{Any of "Top", "Bottom", "Left", "Right".}

\item{thick}{Which of those sides should be thick.}

\item{thin_colour}{Colour of the thin sides.}

\item{thick_colour}{Colour of the thick sides.}
}
\value{
A list with \code{border}, \code{borderColour} and \code{borderStyle}.
}
\description{
openxlsx assigns colours and styles to sides in the order the side names
appear in the \code{border} string, so all three values are built together.
}
\keyword{internal}
