% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{ck_gsum}
\alias{ck_gsum}
\title{Group-Wise Sums of a Numeric Vector}
\usage{
ck_gsum(values, g, levels_g)
}
\arguments{
\item{values}{Numeric vector to sum.}

\item{g}{Group vector (character).}

\item{levels_g}{Group levels to report, in order.}
}
\value{
A numeric vector, one element per level of \code{levels_g}.
}
\description{
Group-Wise Sums of a Numeric Vector
}
\keyword{internal}
