% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{ck_loa_columns}
\alias{ck_loa_columns}
\title{Columns analysistools::create_analysis() Understands}
\usage{
ck_loa_columns()
}
\description{
Columns analysistools::create_analysis() Understands
}
\keyword{internal}
