% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pipeline.R
\name{ck_require_pkg}
\alias{ck_require_pkg}
\title{Require a Suggested Package}
\usage{
ck_require_pkg(pkg, repo = NULL)
}
\arguments{
\item{pkg}{Package name.}

\item{repo}{Optional GitHub org/repo used in the install hint.}
}
\description{
Require a Suggested Package
}
\keyword{internal}
