% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_as_number}
\alias{loa_as_number}
\title{Coerce a Spreadsheet Cell to a Number}
\usage{
loa_as_number(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A numeric vector; unparseable values are \code{NA}.
}
\description{
Coerce a Spreadsheet Cell to a Number
}
\keyword{internal}
