% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_blank_rows}
\alias{loa_blank_rows}
\title{Is Every Cell of This Row Empty}
\usage{
loa_blank_rows(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A logical vector, one element per row.
}
\description{
Is Every Cell of This Row Empty
}
\keyword{internal}
