% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_collect}
\alias{loa_collect}
\title{Bind a List of Problem Tables}
\usage{
loa_collect(found)
}
\arguments{
\item{found}{A list of problems data frames.}
}
\value{
One problems data frame, errors first.
}
\description{
Bind a List of Problem Tables
}
\keyword{internal}
