% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_combination_list}
\alias{loa_combination_list}
\title{Build the count_combinations List from its Sheet}
\usage{
loa_combination_list(sheet)
}
\arguments{
\item{sheet}{The \code{count_combinations} data frame.}
}
\value{
A named list of named character vectors.
}
\description{
Build the count_combinations List from its Sheet
}
\keyword{internal}
