% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_coverage_problems}
\alias{loa_coverage_problems}
\title{Turn a Coverage Table into Problems}
\usage{
loa_coverage_problems(coverage)
}
\arguments{
\item{coverage}{The output of \code{\link{loa_variable_coverage}}.}
}
\value{
A problems data frame.
}
\description{
Turn a Coverage Table into Problems
}
\keyword{internal}
