% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_exclude_choices}
\alias{loa_exclude_choices}
\title{Build the exclude_choices Vector from its Sheet}
\usage{
loa_exclude_choices(sheet)
}
\arguments{
\item{sheet}{The \code{exclude_choices} data frame.}
}
\value{
A character vector.
}
\description{
Build the exclude_choices Vector from its Sheet
}
\keyword{internal}
