% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_known_analysis_types}
\alias{loa_known_analysis_types}
\title{\code{analysis_type} Values the Pipeline Accepts}
\usage{
loa_known_analysis_types()
}
\value{
A character vector.
}
\description{
\code{analysis_type} Values the Pipeline Accepts
}
