% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_no_problems}
\alias{loa_no_problems}
\title{An Empty Problems Table}
\usage{
loa_no_problems()
}
\value{
A zero-row data frame with the problems columns.
}
\description{
An Empty Problems Table
}
\keyword{internal}
