% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loa.R
\name{loa_rename_values}
\alias{loa_rename_values}
\title{Apply a Rename Map to a Vector of Variable Names}
\usage{
loa_rename_values(x, map)
}
\arguments{
\item{x}{Character vector of variable names.}

\item{map}{A named character vector as in \code{\link{apply_rename_map}}.}
}
\value{
\code{x} with mapped names replaced.
}
\description{
Apply a Rename Map to a Vector of Variable Names
}
\keyword{internal}
